First commit of stft-up tools
Contains code for:

**Single Channel Speech Enhancement**

This includes the IMCRA noise variance estimator

    I. Cohen, "Noise Spectrum Estimation in Adverse Environments: Improved Minima Controlled
    Recursive Averaging", in IEEE Trans. on Speech and Audio Processing, Vol 11 (5), pp 1063-6676,
    2003

along with MMSE-LSA and MMSE-MFCC (attained through STFT-UP) estimators.

**MFCC feature extraction with uncertainty propagation**

This includes the STFT-UP implementation for a MFCC feature extraction with delta, acceleration features and cepstral mean subtraction (see example1.m). With aid of the voicebox toolbox the code also allows to replace HCopy with a custom matlab function (see example2.m). This can be used to reproduce the experiments in

    R. F. Astudillo, "Integration of short-time Fourier domain speech enhancement and observation
    uncertainty techniques for robust automatic speech recognition", Ph.D. dissertation, Technische
    Universitaet Berlin, 2010.